// src/app/models/RegistrationRequestModel.ts
export class RegistrationRequestModel {
    companyUserName: string;
    companyPassword: string;
    companyEmailAddress: string;
    companyName: string;
    companyType: string;
    yearOfEstablishment: string;
    companySize: number;
    companyAddress: string;
    companyRepContactName: string;
    companyRepPhoneNumber: number;
    companyRepEmailAddress: string;
    interestedIn: string;
    isApproved: boolean;
}